Convert buffer code to native bytes.Buffer type
This might have actually slowed things down a tad, but it's barely
measurable, while VMpeak showed 3.8G (versus 3.7G for previous
versions).

But it's one less use of Malloc.

Might replace building a buffer with building a string, to avoid the
overhead of converting the resulting buffer to a string.

There are still a few more Malloc's left, though....