Use 'new' instead of 'Malloc' for Lisp objects
This is a quick-and-dirty way to remove the final Malloc() calls.

The result has the same bug(s) preventing passing the basic tests.

The performance looks okay, maybe better than the last version, though
that could be jitter (it ran a little slower once, then a few seconds
faster the second time, and I ran it only once the previous time).

VMpeak back down to 3.7G. That's still rather...insane....