Nativize more code -- PASSES TESTS
Cleaned out the autogenerated header stuff that wasn't needed.

Got rid of more 'unsafe' stuff. This is probably why the tests now
pass -- the garbage collector (GC) might not have been finding stuff
that was allocated, and freeing it, causing problems later on?

Flush stderr and stdout on assert failures.